Revert "Use feature flags on the imports"
This reverts commit 5988d8a221c4f8bb8b9dd095fc703d1e5e99d7ef.

Revert "Guard common types by feature flags"

This reverts commit f12cdc650765da455bf29f1eca0205662a039267.

Revert "Use common types"

This reverts commit a012d1ce2b0ff83039a8ecd7097d66fa96d62b1a.

Revert "Add common types between v14 and v15"

This reverts commit 718c9d39e9c53c5ba51ed249595f8263b8604605.
